@@ -3,13 +3,15 @@
     "description": "MultiFlexi web site",
     "type": "project",
     "require": {
-        "php": "^8.0|^8.1|^8.2",
+        "php": ">=8.1",
         "vitexsoftware/ease-twbootstrap5": "dev-main",
         "vitexsoftware/ease-html-widgets": "dev-main",
         "vitexsoftware/ease-fluentpdo": "dev-main",
         "datatables/datatables": "dev-master",
         "rcubitto/json-pretty": "dev-master",
-        "vitexsoftware/ease-core-bricks": "dev-master"
+        "vitexsoftware/ease-core-bricks": "dev-master",
+        "vitexsoftware/ease-bootstrap5-widgets": "^1.4",
+        "vitexsoftware/multiflexi-core": "dev-main"
     },
     "license": "MIT",
     "authors": [
@@ -37,7 +39,8 @@
         "phpstan/phpstan": "*",
         "friendsofphp/php-cs-fixer": "^3.61",
         "ergebnis/composer-normalize": "^2.43",
-        "ergebnis/php-cs-fixer-config": "^6.34"
+        "ergebnis/php-cs-fixer-config": "^6.34",
+        "robmorgan/phinx": "0.x-dev"
     },
     "config": {
         "allow-plugins": {
